chore: consolidate getExpectedGasLimit
