+++ packages/state-transition/src/util/gloas.ts
@@ -74,27 +74,31 @@ export function isActiveBuilder(builder: gloas.Builder, finalizedEpoch: Epoch): 
 }
 
 /**
+ * Compute the gas limit that satisfies the EIP-1559 adjustment rule from `parentGasLimit`,
+ * clamping `targetGasLimit` into the allowed window of `±max(parentGasLimit / 1024, 1) - 1`.
+ *
+ * From https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1559.md
+ */
+export function getExpectedGasLimit(parentGasLimit: number, targetGasLimit: number): number {
+  const maxGasLimitDifference = Math.max(Math.floor(parentGasLimit / 1024), 1) - 1;
+
+  if (targetGasLimit > parentGasLimit) {
+    const gasDiff = targetGasLimit - parentGasLimit;
+    return parentGasLimit + Math.min(gasDiff, maxGasLimitDifference);
+  }
+
+  const gasDiff = parentGasLimit - targetGasLimit;
+  return parentGasLimit - Math.min(gasDiff, maxGasLimitDifference);
+}
+
+/**
  * Check if `gasLimit` is compatible with `targetGasLimit` under the EIP-1559 transition rule
  * from `parentGasLimit`. The bid must hit `targetGasLimit` when the target is within one
  * adjustment step of the parent, otherwise it must hit the clamped boundary.
  * Spec: https://github.com/ethereum/consensus-specs/blob/v1.7.0-alpha.8/specs/gloas/builder.md#new-is_gas_limit_target_compatible
  */
-export function isGasLimitTargetCompatible(
-  parentGasLimit: number,
-  gasLimit: number,
-  targetGasLimit: number
-): boolean {
-  const maxGasLimitDifference = Math.max(Math.floor(parentGasLimit / 1024), 1) - 1;
-  const minGasLimit = parentGasLimit - maxGasLimitDifference;
-  const maxGasLimit = parentGasLimit + maxGasLimitDifference;
-
-  if (targetGasLimit >= minGasLimit && targetGasLimit <= maxGasLimit) {
-    return gasLimit === targetGasLimit;
-  }
-  if (targetGasLimit > maxGasLimit) {
-    return gasLimit === maxGasLimit;
-  }
-  return gasLimit === minGasLimit;
+export function isGasLimitTargetCompatible(parentGasLimit: number, gasLimit: number, targetGasLimit: number): boolean {
+  return gasLimit === getExpectedGasLimit(parentGasLimit, targetGasLimit);
 }
 
 /**
